junk removed
