--- src/components/FAQ/index.module.scss
@@ -1,104 +1,0 @@
-@import '~@santiment-network/ui/mixins';
-
-.wrapper {
-  position: relative;
-
-  &::before {
-    content: '';
-    display: block;
-    position: absolute;
-    width: 100vw;
-    top: -180px;
-    height: 500px;
-    z-index: -1;
-    background: url('./bg.svg') no-repeat 100% 0;
-    background-size: 30%;
-  }
-}
-
-.questions {
-  margin: 50px auto 0;
-  max-width: 720px;
-}
-
-.question {
-  background: #fff;
-  border: 1px solid var(--porcelain);
-  border-radius: 4px;
-  padding: 18px 24px;
-  cursor: pointer;
-  margin-bottom: 16px;
-  user-select: none;
-
-  &:global(.opened) {
-    box-shadow: 0 20px 50px rgba(24, 27, 43, 0.08);
-  }
-
-  &__top {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-  }
-
-  &__arrow {
-    fill: var(--waterloo);
-    min-width: 15px;
-
-    :global(.opened) & {
-      fill: var(--dodger-blue);
-      transform: rotate(180deg);
-    }
-  }
-
-  &__text {
-    @include text('body-2');
-
-    :global(.opened) & {
-      color: var(--dodger-blue);
-    }
-  }
-
-  &__answer {
-    display: none;
-    margin-top: 18px;
-
-    & * {
-      font-size: 14px;
-    }
-
-    :global(.opened) & {
-      display: block;
-    }
-  }
-}
-
-.heading {
-  margin: 30px 0 20px;
-  font-size: 18px;
-}
-
-.list {
-  padding-left: 20px;
-  margin-bottom: 20px;
-
-  &__item {
-    list-style: disc;
-  }
-}
-
-.text {
-  margin-bottom: 20px;
-
-  a {
-    color: var(--dodger-blue);
-
-    &:hover {
-      color: var(--dodger-blue-hover);
-      text-decoration: underline;
-    }
-  }
-
-  &:last-child {
-    margin: 0;
-  }
-}
